requirements.txt
lxml==5.1.0
PyYAML==6.0.1
